% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{apply_class_power_summary}
\alias{apply_class_power_summary}
\title{Generate html table with power summary}
\usage{
apply_class_power_summary(df)
}
\arguments{
\item{df}{input dataframe}
}
\description{
Generate html table with power summary
}
