% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-00-csv-to-html.R
\name{build_element_apply}
\alias{build_element_apply}
\title{Convert dataframe columns into html elements, then reduce to single string}
\usage{
build_element_apply(df, pre, post, df.names = names(df),
  skipEmpty = TRUE, collapse = "")
}
\arguments{
\item{df}{input dataframe}

\item{pre}{Dataframe with string columns and same names as df. Each column in
pre contains a string value to append before the corresponding
column in df. Includes a Body column, to be appended before the
aggregated output.}

\item{post}{Dataframe with string columns and same names as df. Each column in
post contains a string value to append before the corresponding
column in df. Includes a Body column, to be appended after the
aggregated output.}

\item{df.names}{names of columns to use}

\item{skipEmpty}{If True, null values of str will be converted to ''.}
}
\value{
character
}
\description{
Convert dataframe columns into html elements, then reduce to single string
}
\examples{
  df_pre = tibble::tibble(x='<x>', y='<y>' ,Body='(')
  df_post = tibble::tibble(x= '<x>', y='</y>', Body=')')
  build_element_apply(tibble::tibble(x=c('hello','world'),y=c('1', '2')), df_pre, df_post)
  
}
