% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_html.R
\name{build_table_apply}
\alias{build_table_apply}
\title{Convert an input dataframe of string columns into an HTML table}
\usage{
build_table_apply(
  df,
  df.names = names(df),
  tableClass = NULL,
  skipHeader = FALSE
)
}
\arguments{
\item{df}{Dataframe with string columns, used as input.}

\item{df.names}{List of dataframe columns to include in output.}

\item{tableClass}{value of class html attribute of table. Use to identify
the table, apply CSS format.}

\item{skipHeader}{If FALSE, add df.names as column names in output table.
Otherwise, use first row as column names.}
}
\value{
A string with an HTML table
}
\description{
Convert an input dataframe of string columns into an HTML table
}
\examples{
build_table_apply(tibble::tibble(x=c('hello','world'),y=c('1', '2')) )

}
