% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-00-csv-to-html.R
\name{buildElement}
\alias{buildElement}
\title{Append tags before and after elements in a string array}
\usage{
buildElement(s, pre, post, skipEmpty = TRUE)
}
\arguments{
\item{s}{string}

\item{pre}{string}

\item{post}{string}

\item{skipEmpty}{logical}
}
\value{
string
}
\description{
Typically used to convert raw data into html elements
}
\examples{
str_result = buildElement(c('hello','world'),'<<', '>>')

}
