% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{clean_df_feature}
\alias{clean_df_feature}
\title{clean class feature table}
\usage{
clean_df_feature(df_feature_raw, df_feature_tag = NULL)
}
\value{
clean class feature table
}
\description{
clean class feature table
}
