% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{clean_df_power}
\alias{clean_df_power}
\title{Build power block from raw data}
\usage{
clean_df_power(df_power_raw, df_power_tag = NULL, character_sheet = FALSE)
}
\arguments{
\item{character_sheet}{boolean, true if generating table for a character
sheet}
}
\value{
character vector with html power blocks
}
\description{
Build power block from raw data
}
