% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equip.R
\name{extract_equip_table}
\alias{extract_equip_table}
\title{Convenience function to extract subtable from equip_table}
\usage{
extract_equip_table(df_tables_equip, type, subtype = "-")
}
\arguments{
\item{df_tables_equip}{}

\item{type}{}

\item{subtype}{}
}
\description{
Convenience function to extract subtable from equip_table
}
