% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_html.R
\name{fillna_df}
\alias{fillna_df}
\title{Fill NAs in non-factor columns with ''}
\usage{
fillna_df(df)
}
\arguments{
\item{df}{input dataframe}
}
\value{
df with NA values replaced with ''
}
\description{
Fill NAs in non-factor columns with ''
}
