% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_class_build}
\alias{get_class_build}
\title{Extract data for a single class build from df_class}
\usage{
get_class_build(df_class, char_class, char_build)
}
\arguments{
\item{df_class}{}

\item{char_class}{}

\item{char_build}{}
}
\description{
Extract data for a single class build from df_class
}
