% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{get_class_section}
\alias{get_class_section}
\title{Generate html document with class data for all classes}
\usage{
get_class_section(class, build, htm_stat, htm_feature, htm_power)
}
\arguments{
\item{class}{Name of class}

\item{build}{Name of build}

\item{htm_stat}{string with class stats in html form}

\item{htm_feature}{string with class features in html form}

\item{htm_power}{string with class powers in html form}

\item{htm_power_summary}{string with html summary table for class powers}
}
\value{
string with class section in html form
}
\description{
Generate html document with class data for all classes
}
