% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{get_class_stat_trans}
\alias{get_class_stat_trans}
\title{Transpose class stats table, into (key, value) form}
\usage{
get_class_stat_trans(df_class_stat)
}
\arguments{
\item{df_class_stat}{}
}
\description{
Transpose class stats table, into (key, value) form
}
