% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_df_class_stat}
\alias{get_df_class_stat}
\title{Generate clean class ctats table from raw data}
\usage{
get_df_class_stat(df_class_stat_raw)
}
\arguments{
\item{df_class_stat_raw}{class stats table
in raw form}
}
\value{
df_class_stats class stats table in clean form
}
\description{
Generate clean class ctats table from raw data
}
