% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equip.R
\name{get_df_equip}
\alias{get_df_equip}
\title{Generate html tables for equipment from source .csv files}
\usage{
get_df_equip()
}
\value{
List of strings with html tables
}
\description{
Generate html tables for equipment from source .csv files
}
\examples{
df_tables_equip = get_df_equip()
}
