% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_df_feature_from_sheet}
\alias{get_df_feature_from_sheet}
\title{Get class feature table from character sheet input}
\usage{
get_df_feature_from_sheet(l_feature_id, Class = "sheet", Build = "sheet")
}
\value{
class power table
}
\description{
Get class feature table from character sheet input
}
