% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_stats.R
\name{get_df_hp}
\alias{get_df_hp}
\title{Generate HP table for Player Characters}
\usage{
get_df_hp()
}
\value{
DataFrame with HP and Surge Values per level
}
\description{
Generate HP table for Player Characters
}
