% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_stats.R
\name{get_df_hp_beast}
\alias{get_df_hp_beast}
\title{Generate HP table for Beast Companions}
\usage{
get_df_hp_beast()
}
\value{
DataFrame with HP per level
}
\description{
Generate HP table for Beast Companions
}
