% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_stats.R
\name{get_df_hp_html_beast}
\alias{get_df_hp_html_beast}
\title{Generate html with table for Beast Companion HP}
\usage{
get_df_hp_html_beast(df_hp_beast = NULL)
}
\value{
html table
}
\description{
Generate html with table for Beast Companion HP
}
