% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_df_maneuver}
\alias{get_df_maneuver}
\title{Generate dataframe with html text for combat maneuver rules}
\usage{
get_df_maneuver()
}
\arguments{
\item{dir_base}{}
}
\value{
df_class table (Class, Build,
        htm_stat,htm_feature,htm_power_summary,htm_power)
}
\description{
Generate dataframe with html text for combat maneuver rules
}
