% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_df_origin}
\alias{get_df_origin}
\title{Generate dataframe with html text for origin rules}
\usage{
get_df_origin(df_feature_raw = NULL, df_power_raw = NULL)
}
\arguments{
\item{dir_base}{}
}
\value{
df_origin table (Class, Build,
        htm_stat,htm_feature,htm_power_summary,htm_power)
}
\description{
Generate dataframe with html text for origin rules
}
