% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_df_power_from_sheet}
\alias{get_df_power_from_sheet}
\title{Get class power table from character sheet input}
\usage{
get_df_power_from_sheet(l_power_id, Class = "sheet", Build = "sheet")
}
\value{
class power table
}
\description{
Get class power table from character sheet input
}
