% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-00-Weapon-csv-to-html.R,
%   R/1-character_building.R
\name{get_equip_tables}
\alias{get_equip_tables}
\title{Generate html tables for equipment from source .csv files}
\usage{
get_equip_tables(dir_base = here::here())

get_equip_tables(dir_base = here::here())
}
\arguments{
\item{dir_base}{Path of base project directory}

\item{dir_base}{Path of base project directory}
}
\value{
List of strings with html tables

List of strings with html tables
}
\description{
Generate html tables for equipment from source .csv files

Generate html tables for equipment from source .csv files
}
\examples{
df_tables_equip = get_equip_tables(dir_base)
df_tables_equip = get_equip_tables(dir_base)
}
