% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{get_feat_list_htm}
\alias{get_feat_list_htm}
\title{Add feat list div to feat html}
\usage{
get_feat_list_htm(htm_feat)
}
\arguments{
\item{htm_feat}{}
}
\value{

}
\description{
Add feat list div to feat html
}
