% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{get_htm_file}
\alias{get_htm_file}
\title{Add html headers, css data to html string to use in standalone document}
\usage{
get_htm_file(str_htm)
}
\arguments{
\item{htm_str}{}
}
\description{
Add html headers, css data to html string to use in standalone document
}
