% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_html_sheet}
\alias{get_html_sheet}
\title{Generate dataframe with html text for character sheet}
\usage{
get_html_sheet(
  l_feature_id,
  l_power_id,
  l_items = NULL,
  character_name = "Character Sheet"
)
}
\arguments{
\item{dir_base}{}
}
\value{
df_class table (Class, Build,
        htm_stat,htm_feature,htm_power_summary,htm_power)
}
\description{
Generate dataframe with html text for character sheet
}
