% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{get_l_class}
\alias{get_l_class}
\title{Generate dataframe with html text for class rules}
\usage{
get_l_class(dir_base = here::here())
}
\arguments{
\item{dir_base}{}
}
\value{
df_class table (Class, Build,  
        htm_stat,htm_feature,htm_power_summary,htm_power)
}
\description{
Generate dataframe with html text for class rules
}
