% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-00-csv-to-html.R
\name{gsubColwise}
\alias{gsubColwise}
\title{Apply gsub, then as.factor on string/factor columns of dataframe}
\usage{
gsubColwise(df, pattern, replacement)
}
\arguments{
\item{df}{Dataframe with string columns, used as input}

\item{pattern}{regex to apply}

\item{replacement}{replacement for pattern}
}
\value{
A dataframe like df1, with pattern replaced by replacement in columns
}
\description{
Apply gsub, then as.factor on string/factor columns of dataframe
}
