% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{read_df_class_feature}
\alias{read_df_class_feature}
\title{Read class feature data from partial .csv tables}
\usage{
read_df_class_feature(is_class = TRUE)
}
\description{
Read class feature data from partial .csv tables
}
