% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{read_df_class_power}
\alias{read_df_class_power}
\title{#' Read class power data from partial .csv tables}
\usage{
read_df_class_power(is_class = TRUE)
}
\description{
#' Read class power data from partial .csv tables
}
