% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{read_df_origin_stats}
\alias{read_df_origin_stats}
\title{Read origin stats data from partial .csv tables}
\usage{
read_df_origin_stats()
}
\value{

}
\description{
Read origin stats data from partial .csv tables
}
