% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{read_df_power}
\alias{read_df_power}
\title{Read class power data from partial .csv tables}
\usage{
read_df_power(is_class = TRUE)
}
\value{
merged class power table, raw
}
\description{
Read class power data from partial .csv tables
}
