% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_html.R
\name{read_my_csv}
\alias{read_my_csv}
\title{Read and clean csv files}
\usage{
read_my_csv(s, delim = ",", folder = "character_creation")
}
\arguments{
\item{s}{string , name of .csv file in /inst/raw/character_creation,}

\item{delim}{string, single character used to separate csv fields}
}
\value{
Dataframe generated from csv
}
\description{
Read and clean csv files
}
