% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-00-csv-to-html.R
\name{refactor}
\alias{refactor}
\title{Apply factor again on factor columns}
\usage{
refactor(df)
}
\arguments{
\item{df}{df with factor columns}
}
\value{
df with refactorized columns
}
\description{
Apply factor again on factor columns
}
