% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_html.R
\name{trans_df}
\alias{trans_df}
\title{Transpose dataframe, into (key, value) form}
\usage{
trans_df(df_class_stat)
}
\arguments{
\item{df_class_stat}{}
}
\value{

}
\description{
Transpose dataframe, into (key, value) form
}
