% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{write_class_file}
\alias{write_class_file}
\title{Write class data for a single class into an html file}
\usage{
write_class_file(df_class, char_class, char_build, dir_output)
}
\arguments{
\item{df_class}{}

\item{char_class}{}

\item{char_build}{}

\item{dir_output}{}
}
\description{
Write class data for a single class into an html file
}
