% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-Class-csv-to-html.R
\name{write_class_section_file}
\alias{write_class_section_file}
\title{Write class data for all classes into an html file}
\usage{
write_class_section_file(df_class, dir_output)
}
\arguments{
\item{df_class}{}

\item{dir_output}{}
}
\description{
Write class data for all classes into an html file
}
