% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{apply_class_power_htm}
\alias{apply_class_power_htm}
\title{Generate html table for powers}
\usage{
apply_class_power_htm(df, pre, post)
}
\arguments{
\item{df}{input dataframe}

\item{post}{}
}
\description{
Generate html table for powers
}
