% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_html.R
\name{build_element}
\alias{build_element}
\title{Append tags before and after elements in a string array}
\usage{
build_element(s, pre, post, skipEmpty = TRUE)
}
\arguments{
\item{s}{string}

\item{pre}{string}

\item{post}{string}

\item{skipEmpty}{logical}
}
\value{
string
}
\description{
Typically used to convert raw data into html elements
}
\examples{
str_result = build_element(c('hello','world'),'<<', '>>')

}
