% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-00-csv-to-html.R
\name{buildElement}
\alias{buildElement}
\title{Build html element from dataframe}
\usage{
buildElement(str, pre, post, skipEmpty = TRUE)
}
\arguments{
\item{str}{string}

\item{pre}{string}

\item{post}{string}

\item{skipEmpty}{logical}
}
\value{
string
}
\description{
On a dataframe column of strings, 
appends strings before and after each element
}
