% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-01-00-Weapon-csv-to-html.R
\name{get_equip_tables}
\alias{get_equip_tables}
\title{Generate html tables for equipment from source .csv files}
\usage{
get_equip_tables()
}
\value{
List of strings with html tables
}
\description{
Generate html tables for equipment from source .csv files
}
\examples{
df_tables_equip = get_equip_tables()
}
