% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_html.R
\name{get_htm_file}
\alias{get_htm_file}
\title{Add html headers, css data to html string to use in standalone document}
\usage{
get_htm_file(str_htm, css = NULL)
}
\arguments{
\item{htm_str}{}
}
\value{

}
\description{
Add html headers, css data to html string to use in standalone document
}
