% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_class.R
\name{get_power_clean}
\alias{get_power_clean}
\title{Build power block from raw data}
\usage{
get_power_clean(df_power_raw, df_power_tag, character_sheet = FALSE)
}
\arguments{
\item{df_power_raw}{}
}
\value{
character vector with html power blocks
}
\description{
Build power block from raw data
}
