% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-00-csv-to-html.R
\name{read_my_csv}
\alias{read_my_csv}
\title{Read and clean csv files}
\usage{
read_my_csv(s, delim = ",", folder = "character_creation")
}
\arguments{
\item{delim}{}
}
\description{
Read and clean csv files
}
